Rearranged classes of primitives
Split into common, composite, and library specific
