+++ QGL/PulsePrimitives/__init__.py
@@ -1,9 +1,2 @@
-from .common_primitives import overrideDefaults, _memoize, clear_pulse_cache, Id, Z, X90, X90m, Y90, Y90m, Z90, Z90m, Utheta, Xtheta, Ytheta, Ztheta, U90, U, arb_axis_drag,\
-AC, DiAC, CNOT, CNOT_CR, flat_top_gaussian, echoCR, ZX90_CR, MEAS, MeasEcho, BLANK
-from .. import config
-if config.pulse_primitives_lib == 'standard':
-    from .standard_primitives import X, Xm, Y, Ym
-elif config.pulse_primitives_lib == 'all90':
-    from .all90_primitives import X, Xm, Y, Ym
-else:
-    raise Exception("Invalid pulse library")
+from .common_primitives import Id, Z, X90, X90m, Y90, Y90m, Z90, Z90m, Utheta, Xtheta, Ytheta, Ztheta, U90, U, arb_axis_drag, MEAS, BLANK
+from .composite_primitives import AC, DiAC, CNOT, CNOT_CR, flat_top_gaussian, echoCR, ZX90_CR, MeasEcho, X, Y
